Test/tt2 140 sanitise data for integration tests (#132)
* New test data parameters for athena query tests. Timestamps as random ticket ids


* Updated data paths to match test event. Timestamp for random ticket id. Refactored resuable methods


* Correcting nesting of suites


* Resolving PR comment


* Resolving PR comment
